% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here.R
\name{local_here}
\alias{local_here}
\title{temporarily change here() root}
\usage{
local_here(
  new_here,
  chdir = FALSE,
  verbose = FALSE,
  .local_envir = parent.frame()
)
}
\arguments{
\item{new_here}{new temporary here root directory}

\item{chdir}{change working directory also}

\item{verbose}{show here's messages on setting new root}

\item{.local_envir}{the environment to use for scoping}
}
\value{
no return
}
\description{
Changes here() root for current environment
}
\details{
Changes here() to temporarily point to a new directory for active
entironment.
}
\examples{
library(here)

myfun <- function() {

  d <- tempdir()
  dir.create(d)

  local_here(d)

  cat("here() is now: ",here(),"\n")
  stopifnot(d == here())

  # do something

}
}
\author{
Torbjørn Lindahl
}
