% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here.R
\name{local_here}
\alias{local_here}
\title{Temporarily Change Project Root}
\usage{
local_here(
  new_here,
  chdir = FALSE,
  verbose = FALSE,
  .local_envir = parent.frame()
)
}
\arguments{
\item{new_here}{new temporary here root directory}

\item{chdir}{change working directory also}

\item{verbose}{show here's messages on setting new root}

\item{.local_envir}{the environment to use for scoping, see \code{\link[withr:with_dir]{withr::local_dir()}}}
}
\value{
The original here() root
}
\description{
Changes the here() root for current environment
}
\details{
Changes \code{\link[here:here]{here::here()}} to temporarily point to a new directory for the active
entironment. Changes back afterwards.
}
\examples{
library(here)
library(withr)

local({

  d <- local_tempdir()

  cat("here was initially: ", here(), "\n")

  local_here(d)

  cat("here() is now: ",here(),"\n")
  stopifnot(normalizePath(d) == normalizePath(here()))

  # do something that requires here() be elsewhere

})
}
\author{
Torbjørn Lindahl
}
