% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here.R
\name{with_here}
\alias{with_here}
\title{Temporarily Change Project Root}
\usage{
with_here(new_here, expr, chdir = FALSE, verbose = FALSE)
}
\arguments{
\item{new_here}{new temporary here root directory}

\item{expr}{expression to evaluate}

\item{chdir}{change working directory also}

\item{verbose}{show here's messages on setting new root}
}
\value{
the results of the provided R expression
}
\description{
Evaluate an expresion under a temporary here() root
}
\details{
Changes \code{\link[here:here]{here::here()}} to temporarily point to a new directory when
evaluating expression. Changes back afterwards.
}
\examples{
library(here)
library(withr)

d <- local_tempdir()

cat("here() is initially: ", here(), "\n")
with_here(d, cat("but here() is now: ", here(), "\n"))
cat("here() is now again: ", here(), "\n")

}
\author{
Torbjørn Lindahl
}
